feat(section): line 엔티티 embedded 추가(Sections) 구현
