+++ src/main/java/nextstep/subway/section/domain/Section.java
@@ -28,7 +28,7 @@ public class Section {
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn
+	@JoinColumn(name = "line_id")
 	private Line line;
 
 	@ManyToOne(fetch = FetchType.LAZY)
