+++ src/main/java/nextstep/subway/section/domain/Sections.java
@@ -1,0 +1,34 @@
+package nextstep.subway.section.domain;
+
+import java.util.ArrayList;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Set;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.OneToMany;
+
+import lombok.Getter;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.station.domain.Station;
+
+@Getter
+@Embeddable
+public class Sections {
+
+	@OneToMany(mappedBy = "line", cascade = CascadeType.ALL) //, orphanRemoval = true
+	private final List<Section> sections = new ArrayList<>();
+
+	public List<Station> getStations() {
+		Set<Station> result = new LinkedHashSet<>();
+		for (Section section : this.sections) {
+			result.addAll(section.getStations());
+		}
+		return new ArrayList<>(result);
+	}
+
+	public void createSection(Line line, Station upStation, Station downStation, int distance) {
+		this.sections.add(new Section(line, upStation, downStation, distance));
+	}
+}
